+++ PokedexServidor/src/pokedexservidor.h
@@ -33,7 +33,7 @@
 
 /* -- declaraciones de constantes -- */
 
-#define size_header  sizeof(uint16_t) * 4
+#define size_header sizeof(uint32_t) * 4
 
 /* -- configuracion de sockets -- */
